Renamed `scratch` to `jdk17`
