+++ pom.xml
@@ -34,7 +34,7 @@
     <groupId>net.ripe.ipresource</groupId>
     <artifactId>ipresource</artifactId>
     <packaging>jar</packaging>
-    <version>1.53-SNAPSHOT</version>
+    <version>2.0-SNAPSHOT</version>
 
     <parent>
         <groupId>org.sonatype.oss</groupId>
