+++ src/main/java/net/ripe/ipresource/jdk17/Asn.java
@@ -27,7 +27,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package net.ripe.ipresource.scratch;
+package net.ripe.ipresource.jdk17;
 
 import net.ripe.ipresource.IpResourceType;
 import org.apache.commons.lang3.Validate;
